+++ include/event.h
@@ -23,8 +23,8 @@ class Event{
     
     bool activated;
     int xSprites, ySprites;
-    double elapsedTime;
-    double triggerTime;
+    float elapsedTime;
+    float triggerTime;
     
     public:
     
@@ -40,7 +40,7 @@ class Event{
     virtual vector<string> getStrings() {return vector<string>(0);}; 
     virtual void saveEvent(std::ofstream& f){};
     virtual void activate(){};
-    virtual void update(double eT){};
+    virtual void update(float eT){};
     virtual void draw(int p){};
 
 };
@@ -63,7 +63,7 @@ class ChangeMap: public Event
     vector<string> getStrings(); 
     void saveEvent(std::ofstream& f);
     void activate();
-    void update(double eT){};
+    void update(float eT){};
     void draw(int p){};
 };
 
@@ -76,8 +76,8 @@ class TextInteraction: public Event
     unsigned int characterSize;
     int thickness;
     int iText;
-    double margin;
-    double interline;
+    float margin;
+    float interline;
     std::vector<sf::String> text;
     std::vector<sf::Text> textToDraw;
     sf::Font font;
@@ -94,7 +94,7 @@ class TextInteraction: public Event
     vector<string> getStrings(); 
     virtual void saveEvent(std::ofstream& f);
     virtual void activate();
-    void update(double eT);
+    void update(float eT);
     void draw(int p);
 };
